Extact to utils. Add testing. Use for-of.
